serverless-io-sink/hologres/src/code/requirements.txt
psycopg2-binary
schema